add blend

+++ include/liberosion_linux.h
@@ -41,7 +41,7 @@ extern "C" {
 #endif
 
 HALIDE_FUNCTION_ATTRS
-int erosion(struct halide_buffer_t *_src_buffer, int32_t _width, int32_t _height, uint8_t _size, struct halide_buffer_t *_erosion_buffer);
+int erosion(struct halide_buffer_t *_src_buffer, int32_t _width, int32_t _height, int32_t _size, struct halide_buffer_t *_erosion_buffer);
 
 HALIDE_FUNCTION_ATTRS
 int erosion_argv(void **args);
